test: removeing jasmine testing
